[CP] all jittery glyph fixes (#163058)
This is a cherry-pick for all the changes that went into fixing https://github.com/flutter/flutter/issues/149652

It looks like a lot but most of it is testing and refactoring.

- https://github.com/flutter/flutter/pull/161625
- https://github.com/flutter/flutter/pull/162351
- https://github.com/flutter/flutter/pull/162415
- https://github.com/flutter/flutter/pull/162555
- https://github.com/flutter/flutter/pull/162824

All users of Impeller.

Animating text with translations and scales can cause:
  - jitter between glyphs
  - jitter between glyphs and the baseline
  - artifacts when rendering glyphs at non integer scales

Use skia.

Since this edits how text is rendered, the risk is pretty high.  The actual changes are small and there are unit tests for them.  Golden test coverage for cherry-picks is not complete and text rendering golden coverage for android is problematic.

Yes.

The reproduction code in https://github.com/flutter/flutter/issues/149652 is good.