[stable] Cherrypick fix for stretched embedded app on mobile (#164406)
What is the link to the issue this cherry-pick is addressing?

https://github.com/flutter/flutter/issues/159974

Explain this cherry pick in one line that is accessible to most Flutter developers. See [best practices](https://github.com/flutter/flutter/blob/main/docs/releases/Hotfix-Documentation-Best-Practices.md) for examples

Improve performance of CanvasKit by avoiding copying bitmaps when rendering

What is the impact (ex. visual jank on Samsung phones, app crash, cannot ship an iOS app)? Does it impact development (ex. flutter doctor crashes when Android Studio is installed), or the shipping production app (the app crashes on launch)

Improves performance of all apps using CanvasKit. Fixes a bug relating to embedded Flutter apps on mobile.

Is there a workaround for this issue?

No

What is the risk level of this cherry-pick?

Are you confident that your fix is well-tested by automated tests?

What are the steps to validate that this fix works?

Run the example app in the fixed issue on Chrome for Android and verify it works as expected.