[stable] Fixes rendering text rotated 180 degrees (#165080)
cherry-pick of https://github.com/flutter/flutter/pull/165008

Impeller users.

Text that is rotated 180 degrees exactly will render as if it is scaled by {-1, 1} instead of {-1, -1}.

You can rotate text 180.001 degrees instead but the fidelity will be less.

If it's wrong it could break all text rendering.  There are golden tests for it.

Golden tests (which are not run on release branches)

https://github.com/flutter/flutter/issues/164958 has reproduction code